requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
beautifulsoup4==4.6.0
bleach==2.1.2
breadability==0.1.20
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
decorator==4.2.1
docopt==0.6.2
entrypoints==0.2.3
Flask==0.12.2
gunicorn==19.7.1
h5py==2.7.1
html5lib==1.0.1
idna==2.6
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
lxml==4.1.1
MarkupSafe==1.0
mistune==0.8.3
newsapi-python==0.0.2
nltk==3.2.5
numpy==1.12.1
oauthlib==2.0.6
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.1
pytrends==4.3.0
pytz==2018.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
readability-lxml==0.6.2
requests==2.13.0
requests-oauthlib==0.8.0
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
sumy==0.7.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
